move memhash to hash/maphash
