+++ compiler/natives/src/runtime/runtime.go
@@ -40,19 +40,11 @@ func (e errorString) Error() string {
 	return "runtime error: " + string(e)
 }
 
-var (
-	_ = jsInitPkg()
-)
-
-func jsInitPkg() *js.Object {
+func init() {
 	jsPkg := js.Global.Get("$packages").Get("github.com/gopherjs/gopherjs/js")
 	js.Global.Set("$jsObjectPtr", jsPkg.Get("Object").Get("ptr"))
 	js.Global.Set("$jsErrorPtr", jsPkg.Get("Error").Get("ptr"))
 	js.Global.Set("$throwRuntimeError", js.InternalObject(throw))
-	return jsPkg
-}
-
-func init() {
 	// avoid dead code elimination
 	var e error
 	e = &TypeAssertionError{}
